# workloads/hydra-sidecar/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Hydra sidecar for token hooks, client management, and bulk sync.",
    "title": "Hydra Sidecar API",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/admin/clients": {
      "post": {
        "description": "Proxies client creation to Hydra Admin API and returns the response enriched with client_secret_hash.\n\nResponse fields:\nclient_secret: Plaintext secret (show to user, NEVER store)\nclient_secret_hash: Hash of secret (store this for sync)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "clients"
        ],
        "summary": "Create OAuth2 client.",
        "operationId": "createClient",
        "parameters": [
          {
            "description": "OAuth2 client configuration (passed through to Hydra)",
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/oAuth2Client"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/clientDataResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "502": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/admin/clients/rotate/{client_id}": {
      "post": {
        "description": "Rotates the client secret and returns the new secret along with its hash.\nOptionally accepts client_secret_expires_at to set expiration for the new secret.\n\nResponse fields:\nclient_secret: New plaintext secret (show to user, NEVER store)\nclient_secret_hash: Hash of new secret (update stored value)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "clients"
        ],
        "summary": "Rotate client secret.",
        "operationId": "rotateClient",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ClientID",
            "description": "Client ID",
            "name": "client_id",
            "in": "path",
            "required": true
          },
          {
            "description": "Optional rotation settings",
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/rotateClientRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/clientDataResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          },
          "502": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/admin/clients/{client_id}": {
      "get": {
        "description": "Returns client details from Hydra (passthrough). Note: client_secret is never returned by Hydra.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "clients"
        ],
        "summary": "Get OAuth2 client.",
        "operationId": "getClient",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ClientID",
            "description": "Client ID",
            "name": "client_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/clientDataResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          },
          "502": {
            "$ref": "#/responses/errorResponse"
          }
        }
      },
      "delete": {
        "description": "Deletes a client from Hydra by client_id.",
        "tags": [
          "clients"
        ],
        "summary": "Delete OAuth2 client.",
        "operationId": "deleteClient",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ClientID",
            "description": "Client ID",
            "name": "client_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/noContent"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          },
          "502": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
...
    "/sync/clients": {
      "post": {
        "description": "Performs full reconciliation of clients - creates new, updates existing, deletes removed.\n\nRequest field behavior:\nclient_secret: Must contain the stored hash (from client_secret_hash in creation response)\nclient_secret_hash: Ignored (use client_secret for the hash)",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "clients"
        ],
        "summary": "Bulk sync OAuth2 clients.",
        "operationId": "syncClients",
        "parameters": [
          {
            "description": "Clients to sync (client_secret must contain the stored hash)",
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/syncClientsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/syncResultResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/token-hook": {
      "post": {
        "description": "Called by Hydra during token issuance to inject client metadata into JWT claims.\nRejects expired clients with 403 Forbidden.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "hooks"
        ],
        "summary": "Token hook for JWT claim injection.",
        "operationId": "tokenHook",
        "parameters": [
          {
            "description": "Token hook request from Hydra",
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/tokenHookRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/tokenHookResponseWrapper"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "403": {
            "$ref": "#/responses/tokenHookErrorResponseWrapper"
          }
        }
      }
    }
  },
  "definitions": {
...
    "clientData": {
      "description": "Used for:\nPOST /admin/clients response (client_secret=plaintext, client_secret_hash=hash)\nPOST /admin/clients/rotate/{id} response (client_secret=new plaintext, client_secret_hash=new hash)\nPOST /sync/clients request array element (client_secret_hash=required hash, client_secret=ignored)",
      "title": "ClientData represents an OAuth2 client with sidecar extensions.",
      "allOf": [
        {
          "$ref": "#/definitions/oAuth2Client"
        },
        {
          "type": "object",
          "properties": {
            "client_secret_hash": {
              "description": "Pre-hashed client secret for storage and sync.\n\nIn responses (create/rotate):\nContains the hash of the plaintext secret - store this value.\n\nIn sync requests:\nRequired. Must contain the stored hash value.\nNote: client_secret is ignored in sync requests (use this field instead).",
              "type": "string",
              "x-go-name": "ClientSecretHash"
            }
          }
        }
      ],
      "x-go-name": "ClientData",
      "x-go-package": "github.com/example/hydra-sidecar"
    },
    "clientResult": {
      "type": "object",
      "title": "ClientResult is the result for a single client in sync.",
      "properties": {
        "client_id": {
          "description": "Client ID",
          "type": "string",
          "x-go-name": "ClientID"
        },
        "error": {
          "description": "Error message if status is \"failed\"",
          "type": "string",
          "x-go-name": "Error"
        },
        "status": {
          "description": "Operation status: \"created\", \"updated\", \"deleted\", or \"failed\"",
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "ClientResult",
      "x-go-package": "github.com/example/hydra-sidecar"
    },
...
    "rotateClientRequest": {
      "type": "object",
      "title": "RotateClientRequest is the optional request body for secret rotation.",
      "properties": {
        "client_secret_expires_at": {
          "description": "Unix timestamp when the new secret expires (0 = never)",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ClientSecretExpiresAt"
        }
      },
      "x-go-name": "RotateClientRequest",
      "x-go-package": "github.com/example/hydra-sidecar"
    },
    "syncClientsRequest": {
      "type": "object",
      "title": "SyncClientsRequest is the request body for bulk client sync.",
      "properties": {
        "clients": {
          "description": "Array of clients to sync.\nEach client must have client_secret_hash set to the stored hash value.\nThe client_secret field is ignored (use client_secret_hash instead).",
          "type": "array",
          "items": {
            "$ref": "#/definitions/clientData"
          },
          "x-go-name": "Clients"
        }
      },
      "x-go-name": "SyncClientsRequest",
      "x-go-package": "github.com/example/hydra-sidecar"
    },
    "syncResult": {
      "type": "object",
      "title": "SyncResult is the response from bulk client sync.",
      "properties": {
        "created_count": {
          "description": "Number of clients created",
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreatedCount"
        },
        "deleted_count": {
          "description": "Number of clients deleted",
          "type": "integer",
          "format": "int64",
          "x-go-name": "DeletedCount"
        },
        "failed_count": {
          "description": "Number of operations that failed",
          "type": "integer",
          "format": "int64",
          "x-go-name": "FailedCount"
        },
        "results": {
          "description": "Per-client operation results",
          "type": "array",
          "items": {
            "$ref": "#/definitions/clientResult"
          },
          "x-go-name": "Results"
        },
        "updated_count": {
          "description": "Number of clients updated",
          "type": "integer",
          "format": "int64",
          "x-go-name": "UpdatedCount"
        }
      },
      "x-go-name": "SyncResult",
      "x-go-package": "github.com/example/hydra-sidecar"
    },
    "tokenHookErrorResponse": {
      "type": "object",
      "title": "TokenHookErrorResponse represents an error response to Hydra token hook.",
      "properties": {
        "error": {
          "description": "Error code (e.g., \"access_denied\")",
          "type": "string",
          "x-go-name": "Error"
        },
        "error_description": {
          "description": "Human-readable error description",
          "type": "string",
          "x-go-name": "ErrorDescription"
        }
      },
      "x-go-name": "TokenHookErrorResponse",
      "x-go-package": "github.com/example/hydra-sidecar"
    },
    "tokenHookRequest": {
      "type": "object",
      "title": "TokenHookRequest represents the incoming request from Hydra token hook.",
      "properties": {
        "request": {
          "type": "object",
          "properties": {
            "client_id": {
              "type": "string",
              "x-go-name": "ClientID"
            },
            "granted_scopes": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "x-go-name": "Scopes"
            }
          },
          "x-go-name": "Request"
        },
        "session": {
          "type": "object",
          "properties": {
            "client_id": {
              "type": "string",
              "x-go-name": "ClientID"
            }
          },
          "x-go-name": "Session"
        }
      },
      "x-go-name": "TokenHookRequest",
      "x-go-package": "github.com/example/hydra-sidecar"
    },
    "tokenHookResponse": {
      "type": "object",
      "title": "TokenHookResponse represents the response to Hydra token hook.",
      "properties": {
        "session": {
          "type": "object",
          "properties": {
            "access_token": {
              "type": "object",
              "additionalProperties": {},
              "x-go-name": "AccessToken"
            }
          },
          "x-go-name": "Session"
        }
      },
      "x-go-name": "TokenHookResponse",
      "x-go-package": "github.com/example/hydra-sidecar"
    }
  },
  "responses": {
    "clientDataResponse": {
      "description": "ClientDataResponse wraps ClientData for swagger response.",
      "schema": {
        "$ref": "#/definitions/clientData"
      }
    },
    "errorResponse": {
      "description": "ErrorResponse represents an error response.",
      "schema": {
        "type": "object",
        "properties": {
          "error": {
            "description": "Error message",
            "type": "string",
            "x-go-name": "Error"
          }
        }
      }
    },
...
    "noContent": {
      "description": "NoContentResponse represents a 204 No Content response."
    },
    "syncResultResponse": {
      "description": "SyncResultResponse wraps SyncResult for swagger response.",
      "schema": {
        "$ref": "#/definitions/syncResult"
      }
    },
    "tokenHookErrorResponseWrapper": {
      "description": "TokenHookErrorResponseWrapper wraps TokenHookErrorResponse for swagger.",
      "schema": {
        "$ref": "#/definitions/tokenHookErrorResponse"
      }
    },
    "tokenHookResponseWrapper": {
      "description": "TokenHookResponseWrapper wraps TokenHookResponse for swagger.",
      "schema": {
        "$ref": "#/definitions/tokenHookResponse"
      }
    }
  }
}